change working directory when deleting the worktree which is currently pwd

+++ git-tree-zsh.plugin.zsh
@@ -65,12 +65,22 @@ git-tree() {
             gt_help
         fi
     elif [ "$1" = "remove" ] || [ "$1" = "-d" ] || [ "$1" = "-D" ]; then
-        local root worktrees branches selection
+        local root worktrees branches selection currentdir worktreepath
+        currentdir=$(pwd) &&
         root=$(git worktree list | head -1 | awk '{print $1}') &&
         worktrees=$(basename $(git worktree list | head -1 | awk '{print $1}'))-worktrees &&
         branches=$(git worktree list) &&
         selection=$(echo "$branches" |rev|awk '{print $1}'|cut -b 2-|rev|cut -b 2- | fzf-tmux -p 80% --no-sort --ansi -0 --height=50% --preview-window 70% --preview 'if output=$(git log origin/{} --color --graph --oneline &>/dev/null); then git fetch &>/dev/null && git log origin/{} --color --graph --oneline; else echo "No log"; fi' +m) &&
-        git worktree remove $(echo "$selection" | head -1 | awk '{print $1}') --force;
+        worktreepath=$(echo "$root/../$worktrees/$selection")
+
+        if [[ ! -z "$selection" ]]; then
+            echo $currentdir
+            echo $worktreepath
+            if [[ "$currentdir" -ef "$worktreepath" ]]; then
+                cd $root
+            fi
+            git worktree remove $(echo "$selection" | head -1 | awk '{print $1}') --force;
+        fi
     else
         gt_help
     fi
